Fix cross gcc ppc 11.2.0 builds wrt format warnings
As described in https://gcc.gnu.org/bugzilla/show_bug.cgi?id=91148 some
formatting error are raised, but are not failing the build.
GH seems to be treating stderr output (maybe containing ERROR) as error and
marks the build as failed.

Add --disable-build-format-warnings to ppc builds to silence this, this is
harmless for us.

Signed-off-by: Marc Poulhiès <dkm@kataplop.net>